+++ CHANGELOG.md
@@ -15,6 +15,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Added option to track variant subscriptions to Tracker. remp/web#2404
 - Added Mailer's segment "Everyone" which lists all subscribers known to Mailer. remp/crm#2973
   - This segment should ideally replace `all_users` provided by CRM and effectively serve as a default. Mailer still filters users based on their newsletter subscription to the email they're receiving.
+- URL parser generator's segment is now optional. remp/crm#2973
+  - If not provided, Mailer's segment with subscribers of selected mail type is used as a default.
 
 ## Archive